Fixed READ_USN_JOURNAL_DATA.FirstUsn type per #215.

+++ PInvoke/Kernel32/WinIOCtl.Structs.cs
@@ -2118,7 +2118,7 @@ namespace Vanara.PInvoke
 			/// Buffer of Change Journal Records.
 			/// </para>
 			/// </summary>
-			public int StartUsn;
+			public long StartUsn;
 
 			/// <summary>
 			/// <para>
@@ -2362,7 +2362,7 @@ namespace Vanara.PInvoke
 			/// Buffer of Change Journal Records.
 			/// </para>
 			/// </summary>
-			public int StartUsn;
+			public long StartUsn;
 
 			/// <summary>
 			/// <para>
